refactor(artifactverifier): share detached artifact verification across OPA and manifest loader
